[rf] Refactor various route functions
Refactor the different functions in the main route and also decorate the functions with the auth.login_required decorator so as to all for only authenticated users

+++ app/main/routes.py
@@ -3,15 +3,17 @@ This script contains the routes for the different API methods.
 
 This handles the overall routing of the application.
 """
-from flask import abort, jsonify, request
+from flask import g, jsonify, request
 
 from . import main
 from app import errors
+from app.auth.routes import auth
 from app.decorators import json, paginate
 from app.models import BucketList, Items, User
 
 
-@main.route('/bucketlists', methods=['GET'])
+@main.route('/bucketlists/', methods=['GET'])
+@auth.login_required
 @paginate()
 def get_bucketlists():
     """
@@ -20,30 +22,15 @@ def get_bucketlists():
     Displays a json of all the created BucketLists and the various items
     associated with them.
     """
-    if not BucketList.query.all():
-        return errors.not_found("You currently have no BucketList saved.")
-
-    bucketlists = []
-
-    user = User.query.filter_by(username='andela').first().user_id
-
     if request.args.get('q'):
-        try:
-            result = BucketList.query.filter_by(created_by=user).filter(
-                BucketList.name.contains(request.args.get('q'))).first()
-        except:
-            return errors.not_found(
-                "No BucketList exist with the name {}"
-                .format(request.args.get('q'))
-            )
-        finally:
-            bucketlists = result.to_json()
-            return jsonify({'bucketlist': bucketlists}), 200
+        return BucketList.query.filter_by(created_by=g.user.user_id).filter(
+            BucketList.name.contains(request.args.get('q')))
     else:
-        return BucketList.query.filter_by(created_by=user)
+        return BucketList.query.filter_by(created_by=g.user.user_id)
 
 
 @main.route('/bucketlists/<int:list_id>', methods=['GET'])
+@auth.login_required
 @json
 def get_bucketlist(list_id):
     """
@@ -52,13 +39,14 @@ def get_bucketlist(list_id):
     Return a json of all the information as regards a particular BucketList.
     """
     bucketlist = BucketList.query.filter_by(bucketlist_id=list_id).first()
-    if not bucketlist:
+    if not bucketlist or bucketlist.created_by != g.user.user_id:
         return errors.not_found("The BucketList with the id: {0} doesn't"
                                 " exist.".format(list_id))
     return bucketlist, 200
 
 
-@main.route('/bucketlists/', methods=['POST'])
+@main.route('/bucketlists/', methods=['POST'], strict_slashes=False)
+@auth.login_required
 @json
 def create_bucketlist():
     """
@@ -70,7 +58,7 @@ def create_bucketlist():
         return errors.bad_request("Only JSON object is accepted. Please "
                                   "confirm that the key 'name' exists.")
 
-    user = User.query.filter_by(username='andela').first().user_id
+    user = User.query.filter_by(username=g.user.username).first().user_id
     try:
         bucketlist = BucketList(name=request.json.get('name'), created_by=user)
         bucketlist.save()
@@ -81,6 +69,7 @@ def create_bucketlist():
 
 
 @main.route('/bucketlists/<int:list_id>', methods=['PUT'])
+@auth.login_required
 @json
 def update_bucketlist(list_id):
     """
@@ -88,16 +77,17 @@ def update_bucketlist(list_id):
 
     Update a BucketList name.
     """
-    bucketlist = BucketList.query.get_or_404(list_id)
+    bucketlist = BucketList.query.filter_by(bucketlist_id=list_id).first()
+
+    if not bucketlist or bucketlist.created_by != g.user.user_id:
+        return errors.not_found("The BucketList with the id: {0} doesn't"
+                                " exist.".format(list_id))
 
     if not request.json:
-        abort(400)
+        return errors.bad_request("No JSON file detected.")
 
     if 'done' not in request.json or 'name' not in request.json:
-        abort(400)
-
-    if type(request.json['name']) != unicode:
-        abort(400)
+        return errors.bad_request("The keys 'done' and 'name' not in the JSON")
 
     bucketlist.name = request.json.get('name')
     bucketlist.save()
@@ -105,22 +95,27 @@ def update_bucketlist(list_id):
 
 
 @main.route('/bucketlists/<int:list_id>', methods=['DELETE'])
+@auth.login_required
 def delete_bucketlist(list_id):
     """
     Delete a BucketList.
 
     Deletes a BucketList and all items associated with it.
     """
-    bucketlist = BucketList.query.get_or_404(list_id)
+    bucketlist = BucketList.query.filter_by(bucketlist_id=list_id).first()
 
-    if not bucketlist:
-        abort(404)
+    if not bucketlist or bucketlist.created_by != g.user.user_id:
+        return errors.not_found("The BucketList with the id: {0} doesn't"
+                                " exist.".format(list_id))
 
     bucketlist.delete()
     return jsonify({'Delete': True}), 200
 
 
-@main.route('/bucketlists/<int:list_id>/items/', methods=['POST'])
+@main.route(
+    '/bucketlists/<int:list_id>/items/', methods=['POST'], strict_slashes=False
+)
+@auth.login_required
 @json
 def add_bucketlist_item(list_id):
     """
@@ -129,9 +124,14 @@ def add_bucketlist_item(list_id):
     This function adds a new item to a BucketList. It gets the name and done
     keys from the json supplied and saves to the database.
     """
-    bucketlist = BucketList.query.get_or_404(list_id)
+    bucketlist = BucketList.query.filter_by(bucketlist_id=list_id).first()
+
+    if not bucketlist or bucketlist.created_by != g.user.user_id:
+        return errors.not_found("The BucketList with the id: {0} doesn't"
+                                " exist.".format(list_id))
+
     if not request.json or 'name' not in request.json:
-        abort(400)
+        return errors.bad_request("No JSON file detected.")
 
     item = Items().from_json(request.json)
     item.bucketlist_id = bucketlist.bucketlist_id
@@ -142,6 +142,7 @@ def add_bucketlist_item(list_id):
 @main.route(
     '/bucketlists/<int:list_id>/items/<int:item_id>', methods=['PUT']
 )
+@auth.login_required
 @json
 def update_bucketlist_item(list_id, item_id):
     """
@@ -149,14 +150,19 @@ def update_bucketlist_item(list_id, item_id):
 
     This function updates an item in a BucketList.
     """
-    bucketlist = BucketList.query.get_or_404(list_id)
-    item = Items.query.get_or_404(item_id)
+    bucketlist = BucketList.query.filter_by(bucketlist_id=list_id).first()
 
-    if not request.json:
-        abort(400)
+    if not bucketlist or bucketlist.created_by != g.user.user_id:
+        return errors.not_found("The BucketList with the id: {0} doesn't"
+                                " exist.".format(list_id))
 
-    if 'name' in request.json and type(request.json['name']) != unicode:
-        abort(400)
+    item = Items.query.get(item_id)
+    if not item or (item.bucketlist_id != bucketlist.bucketlist_id):
+        return errors.not_found("The item with the ID: {0} doesn't exist"
+                                .format(item_id))
+
+    if not request.json or (('name' or 'done') not in request.json):
+        return errors.bad_request("No JSON file detected.")
 
     item.name = request.json.get('name')
     item.done = request.json.get('done')
@@ -167,6 +173,7 @@ def update_bucketlist_item(list_id, item_id):
 @main.route(
     '/bucketlists/<int:list_id>/items/<int:item_id>', methods=['DELETE']
 )
+@auth.login_required
 @json
 def delete_bucketlist_item(list_id, item_id):
     """
@@ -174,31 +181,37 @@ def delete_bucketlist_item(list_id, item_id):
 
     This function deletes an item from a BucketList.
     """
-    bucketlist = BucketList.query.get_or_404(list_id)
-    item = Items.query.get_or_404(item_id)
+    bucketlist = BucketList.query.filter_by(bucketlist_id=list_id).first()
 
-    if item.bucketlist_id != bucketlist.bucketlist_id:
-        abort(500)
+    if not bucketlist or bucketlist.created_by != g.user.user_id:
+        return errors.not_found("The BucketList with the id: {0} doesn't"
+                                " exist.".format(list_id))
+
+    item = Items.query.filter_by(item_id=item_id).first()
+
+    if not item or (item.bucketlist_id != bucketlist.bucketlist_id):
+        return errors.not_found("The item with the ID: {0} doesn't exist"
+                                .format(item_id))
 
     item.delete()
     return bucketlist, 200
 
 
-@main.route('/login', methods=["POST", "GET"])
+@main.route('/login', methods=["POST"], strict_slashes=False)
 def login2():
     """
     Secure redundant route.
 
     Protect the route: /login from being accessed by the end user.
     """
-    return jsonify({"message": "Page not allowed."}), 404
+    return errors.not_found("The page doesn't exist")
 
 
-@main.route('/register')
+@main.route('/register', methods=["POST"], strict_slashes=False)
 def register_user2():
     """
     Secure redundant route.
 
     Protect the route: /register from being accessed by the end user.
     """
-    return jsonify({"message": "Page not allowed."}), 404
+    return errors.not_found("The Page doesn't exist")
